Integrate trees23 into stdlib.
If you run Pure with --enable=trees23, then trees23 will be used (under
the namespace "baltree") as the default balanced tree. If avltrees
is used, it will use the namespace "avl".

If you run Pure with --disable=trees23 (currently the default), then
avltrees will be used as the default and will occupy namespace
"baltree". If trees23 is used, it will use the namespace "trees23".