Update testing/log-making system.
This is a bit more complex than I'd like, but it works.

Any new test that cares about the compiler settings (even if only
because that makes a difference to the log output) should have an entry
in index.aux, and also a testXXX.aux file to skip the test when the
compiler settings aren't appropriate. The index.aux entry is used as a
default when running the test, and is also used by the Makefile to
generate the golden log.

The prelude tests work a bit differently. See index.aux and prelude.aux
for examples. If any prelude test fails, `run-tests -f` will re-run all
of the prelude tests. It seemed too complicated to try anything smarter.