Port reverse_onto to strings.pure.
Putting a ::type modifier onto the base argument will cause several
tests with infinite streams (test023, test024, test080) to fail.

So we have the stdlibs instead use a __std__::reverse_onto that
does minimal typechecking on its arguments. Functions should only
use this when they know that the second argument is a proper list
(for example, because they constructed it themselves).

Then the toplevel API is free to present a reverse_onto that
does typecheck its arguments. This returns a string value when
the second argument is a string and the base is coercable into one.