Rename dict.pure's private namespace to "dict".
None of heap.pure, array.pure, and avltrees.pure use a "__" prefix
for their private namespaces.

Note we haven't actually made the contents of the dict namespace
private. Doing so would be tedious but straightforward (all the
definitions for the toplevel namespace should be declared public
and use a leading "::").

If we do that, though, why not also make the *Dict constructors
private? And why not also the *Set and *Bag constructors? And
why not the Heap and Array constructors? Given the presence of
public types and/or type predicates for these, do we still want the
constructors to be private?