Add reverse_onto to lists.
We also define reverse in terms of reverse_onto, as:

  reverse xs@(_:_)   = reverse_onto [] xs;

We don't do any more aggressive optimizing, such as:

  reverse (x:xs)     = reverse_onto [x] xs;

in order to be sure that improper tails (for xs) are always handled by reverse_onto.

Many "tacks" in lists.pure and prelude.pure are really uses of
reverse_onto, but we don't update them until a later commit, because
of a difficulty that emerged with strings.pure.