Add folds, all/any, to baltrees, set, dict.
The simple fold methods in the baltree backends take an additional
preprocessing function to apply to each value before applying the fold
function to it. This lets us reuse some of the backend functions for
several different high-level collections. Note that it doesn't suffice
just to compose the preprocessing function with the fold function, when
dealing with foldl1 and foldr1.

The mhfold functions are most complex, and need to be special-cased.

This is what we expose in the high-level API:

    In set.pure:

     * foldl, foldl1, foldr, foldr1, any, all -- bags yield their values
       one at a time
     * mfoldl, mfoldl1, mfoldr, mfoldr1, many, mall -- sets and bags
       both yield lists of values

    In dict.pure:

     * foldl, foldl1, foldr, foldr1, any, all -- mdicts yield their k=>v
       one at a time
     * mfoldl, mfoldl1, mfoldr, mfoldr1, many, mall -- all dicts yield
       k=>list_of_vs