Optimizations to set relations and operators.
Some operations used more traversals than necessary, and the
intermediate coercions for mixed arguments could be avoided in all but a
few cases.

Note that we're making some assumptions here that may not always be
true. (These assumptions were already being made before these
optimizations.) Speciically, just because m1 is an ordered set, and so
'<' is defined on its elements, and m2 is an ordered set, and so '<' is
defined on its elements, it doesn't follow that '<' will be defined on
elements taken pairwise from m1 and m2. Yet we assume that it will.

If the user of this library wants to deal with such cases, she'll have
to do the appropriate coercions herself. Perhaps this should be
explained in the library documentation?

I have inspected these optimizations carefully, and the existing tests
all pass. But it would be good for other eyes to also review them
carefully.