Add rotate to lists and array.
In an earlier commit (not on this branch) I had implemented for lists
only what is here `rotate 1 xs`. That is equivalent to
`last xs:init xs`, but uses only a single traversal.

Since I wanted to supply a general rotate for arrays, it seemed easies
to just give the list and array operations the same interface. We still
follow an optimized path for lists when n == 1.