Small optimization to some list thunks.
Thunks for +, filter, foldr, map, scanl, cat, zip* now refer to local
auxiliaries.

Cannot optimize in the same way:
    * init (either version)
    * span (reverse-based)
    * take (reverse-based)
    * takewhile (reverse-based)

test061 depends on the definition of foldr, so we update its logfile.
Also, at this point we split off a separate version of that test as
test091.pure, to be run when --enable=list-opt.