Special-case implementation of ==? for arrays and heaps.
To be used when `==` is not defined for all the elements. This is more
efficient than recursing once for `==`, failing, then recursing again
(in C) for `===`. More importantly, it gives better results when `==` is
defined for some but not all of the elements.

To prevent the prelude's general case rule from shadowing these
defintions, though, we need to add special-case checks for these
types to prelude.pure. Since we only have a small number of collections
in the stdlib, this seems tolerable.

If we're going to end up wanting special-case implementations of ==? for
more and more collections though, it might become less tolerable. In
that case, we may want to have a convention where, say, a module
declares the result of ``tuple2 x y`` as belonging to the type
``egalable`` (a better name would be nice) if the module knows how to
provide an implementation of ==? for that x and y. Then the prelude's
general case ==? would just apply when ``tuple2 x y`` is not egalable.

(I call it a "convention", but one might arguably characterize it
as "magic".)

The dynamic extensibility of Pure's types would make this work
correctly. Until module X was loaded, its special-case rules for ==?
would be ignored, and the general rule in the prelude would apply
(if operands of the appropriate type were even constructable before
module X was loaded). After module X was loaded, it would take over
for the operands it wants.

But I'm not sure how much of a performance hit there would be: all those
function calls to check operand types every time ==? is applied. *Plus*
building then discarding all those tuple2s.

But let's cross that bridge if and when we have to. With the scheme presented
here, we do have more operand type-checking, but we at least avoid
building lots of temporary tuple2s.